django-postgres-backend/requirements.txt
asgiref==3.2.7
astroid==2.4.1
autopep8==1.4.4
boto3==1.14.4
botocore==1.17.4
certifi==2020.4.5.2
coverage==5.1
Django==3.0.7
django-cors-headers==3.3.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
fonttools==4.11.0
gunicorn==20.0.4
HarfPy==0.9
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.2
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.42
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.0
uharfbuzz==0.10.0
urllib3==1.25.9
wrapt==1.11.2
